move scripts into `gc_derivatization` subdir
